Moved to node.js server, started working on division
There are now two haxe projects: The server side, which is a simple
nodejs server, written with hxnodejs, and the client side, which is the
actual dice-rolling script.
The build.hxml will build both of these and place them in the correct
directory.
The division operator was added to the parser, but actual division
isn’t done yet.
Fixed issue with the URL expanding every time on reload if the dice
notation contained a /.