Added brackets support, error exceptions and handling for them
If there is an error in parsing or evaluation due to incorrect input, a
typed exception will be thrown. This will be caught by `client/Main`,
and an appropriate error message is shown. For errors that are
internal, the error enum number is also displayed for debugging.
Relatedly, on incorrect input the url is not changed to 1d6, since it
is possible there was a typo, and the url shouldn’t be deleted in that
case.